requirements/tests.txt
virtualenv==15.2.0
ipdb==0.11
coverage==4.5.1
mock==2.0.0
tox==3.0.0

pytest==3.4.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-xdist==1.22.2
